+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ## Fixed
 - Fixed using incorrect unmanaged memory free function in `PagentClent.SendMessage()`.
 - Fixed `PagentAgent.Dispose()` returning before all resources have been freed.
+- Fixed `WindowsOpenSshPipe` starting a new server after `Dispose()`.
 
 ## Removed
 - Removed Qt widgets.